% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{comarcas_tjrs}
\alias{comarcas_tjrs}
\title{Lista de foros, comarcas e códigos númericos do TJRS.}
\format{data_frame com 174 linhas e 4 colunas
\describe{
  \item{orgao}{Comarca}
  \item{origem}{Foro}
  \item{codigo}{Código do foro para número CNJ}
  \item{codigo_themis}{Código do foro para número Themis}
}}
\source{
\url{http://www.tjrs.jus.br/site/processos/numeracao_unica_de_processos_judiciais/}
}
\usage{
comarcas_tjrs
}
\description{
Uma tabela que lista os foros pertencentes a cada comarca junto aos seus respectivos códigos.
}
\keyword{datasets}

