% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawlers.R
\name{cpo_pg_um}
\alias{cpo_pg_um}
\title{Consulta um número de processo no sistema do TJRS}
\usage{
cpo_pg_um(p, path_file, path_log = NULL)
}
\arguments{
\item{p}{Vetor com números de processos.}

\item{path_file}{diretório onde os .html serão salvos}

\item{path_log}{diretório onde o log será salvo. Quando recebe NULL, seu valor padrão,
apenas retorna o log em um objeto do R.}
}
\value{
Uma tabela de log do download
}
\description{
Salva as páginas baixadas a partir de um vetor de números de processos.
}

