% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawlers.R
\name{download_diario_oficial_tjrs}
\alias{download_diario_oficial_tjrs}
\title{Download Diário Oficial}
\usage{
download_diario_oficial_tjrs(...)
}
\arguments{
\item{caderno}{Número inteiro que representa o caderno a ser baixado. Dependendo da época esse número tem interpretações diferentes. Veja as vignettes para mais informações.}

\item{edicao}{Número da edição. A primeira edição disponível é a de número 3424, de agosto de 2006. Aparentemente é contínuo, veja as vignettes para mais informações.}

\item{path}{caminho p/ salvar o pdf}
}
\value{
retorna TRUE se o download foi bem realizado de acordo
com a função verify_download.
}
\description{
Faz download de um diário oficial a partir de sua edição e caderno
}
\note{
atualmente os arquivos são salvos no formato:
tjpr_dje_data_numDiario.pdf dentro da pasta especificada
}

