% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawlers.R
\name{download_pdf_tjrs}
\alias{download_pdf_tjrs}
\title{Download de DJE do TJRS}
\usage{
download_pdf_tjrs(...)
}
\arguments{
\item{caderno}{Número inteiro que representa o caderno a ser baixado. Dependendo da época esse número tem interpretações diferentes. Veja as vignettes para mais informações.}

\item{edicao}{Número da edição. A primeira edição disponível é a de número 3424, de agosto de 2006. Aparentemente é contínuo, veja as vignettes para mais informações.}

\item{path}{caminho p/ salvar o pdf}
}
\description{
Faz a requisição para download do PDF de DJE do TJRS
}

