% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawlers.R
\name{extrair_indice}
\alias{extrair_indice}
\title{Procurar um índice num arquivo de texto}
\usage{
extrair_indice(arquivo_texto)
}
\arguments{
\item{arquivo_texto}{Um arquivo.txt bem formatado.}
}
\value{
Um data_frame com duas colunas: topico e página.
}
\description{
Obtém uma tabela de índice a partir de um arquivo de texto. Assume que os tópicos do sumário estejam separados em linhas.
}

