% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawlers.R
\name{lista_dje_mes}
\alias{lista_dje_mes}
\title{Lista cadernos e links disponíveis}
\usage{
lista_dje_mes(mesano)
}
\arguments{
\item{mesano}{String no formato "mês-ano". Apenas os diários de 06-2006 em diante estão disponíveis.}
}
\value{
Um data_frame em que cada linha representa um caderno. Contém a data de publicação do caderno, o link de download e o nome do caderno.
}
\description{
Fornece uma lista dos diários e links disponíveis de um determinado par "mes-ano".
}

