% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawlers.R
\name{matches_busca_livre_tjrs}
\alias{matches_busca_livre_tjrs}
\title{Rapagem de todas as ocorrências de um texto no DJE's do TJRS}
\usage{
matches_busca_livre_tjrs(word, ed = "")
}
\arguments{
\item{palavra}{string com o texto que será procurado. Não é case-sensitive.}
}
\value{
Retorna um data_frame contendo as edições e as páginas dos matchs encontrados.
}
\description{
Raspa todos os matchs de uma busca em
http://www.tjrs.jus.br/busca/?tb=dj
}
\examples{
matches_busca_livre_tjrs('bolota')
}
\seealso{
Da página não é possível raspar o parágrafo do match, mas é possível raspar exatamente a página em que ocorre o pareamento. Não importa quantos matches tenham ocorrido, ele só te mostra, no máximo, 1000.
}

