% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pdf2text}
\alias{pdf2text}
\alias{pdf2text_system}
\alias{pdf2text_verify}
\title{PDF para texto}
\usage{
pdf2text(...)

pdf2text_system(a, new_file, raw, first_pg, last_pg)

pdf2text_verify(a)
}
\arguments{
\item{a}{arquivo PDF que será transformado}

\item{new_file}{nome do arquivo txt}

\item{raw}{(FALSE)}

\item{first_pg}{(NA) primeira página do arquivo}

\item{last_pg}{(NA) última página do arquivo}

\item{keep_file}{(FALSE)}

\item{overwrite}{(FALSE) você quer apagar o arquivo?}
}
\description{
Função para transformar um pdf em texto.

Função auxiliar para simplificar o código.

Faz as verificações necessárias para o pdf2text
}

