% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crawlers.R
\name{verify_download}
\alias{verify_download}
\title{Verificar download}
\usage{
verify_download(path)
}
\arguments{
\item{path}{caminho do arquivo que acabou de ser baixado}
}
\description{
Verifica se o arquivo foi baixado corretamente.
TODO: adicionar mais verificações. Atualmente é só pelo
tamanho do arquivo.
}

